MyPersonalCleaner/requirements.txt
asgiref==3.3.4
azure-common==1.1.27
azure-core==1.13.0
azure-identity==1.5.0
azure-mgmt-core==1.2.2
azure-mgmt-resource==16.0.0
boto3==1.17.46
botocore==1.20.46
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
et-xmlfile==1.0.1
idna==2.10
isodate==0.6.0
jmespath==0.10.0
msal==1.10.0
msal-extensions==0.3.0
msrest==0.6.21
oauthlib==3.1.0
openpyxl==3.0.7
portalocker==1.7.1
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
